Here it is, project finished. :) (#14)
* Here it is, project finished. :)

* Done. I implemented some logic that won´t allow the user to create a contact if it there are no categories existent. It solves the problem of the program "blowing up" in that moment.

I also considered creating a default "category" object that would basically work as a "null" if the user didnt create any objects yet.
I'm not sure, but i think that way it would enable the user to create contacts with "no category" (default) if no categories were yet created by the user.

However, i decided to simply not allow the user to create any contacts at all. As I though, thats the way the app is supposed to work.
-------------
Also, I made some changes that will print a message if the user tries to update or delete (contacts or categories) and there are none existant. Instead of just "blowing up".

* Ok, problem solved. I created a method "isValidEmail" to validate the email. So instead of the program breaking, now it displays an error message.

* Ok, changed it, it should only accept valid emails now.