Create database structure for Contact phone book
This commit introduces the database structure and related models for the Contact phone book. The structure consists of a "Contacts" table, an "EmailAddresses" table linked with foreign keys to the Contacts table, and a "PhoneNumbers" table also linked to the Contacts table. Each Contact can have multiple associated email addresses and phone numbers.