Implement main functionality for PhoneBook application
Added necessary files and logic to run the core functionality of the PhoneBook application. This includes a user interface for menu selection, contact operations such as view, add, update and delete, as well as a setup for a new service layer. Changes were made in multiple files and some new ones were created, such as ContactController, UserInterface, MenuOptions, and ContactService. Also added necessary configurations in the .csproj file to preserve appsettings.json on output.